Migrate /test/apm_api_integration/tests/custom_dashboards to be deployment-agnostic API tests
