+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/index.ts
@@ -12,5 +12,6 @@ export default function apmApiIntegrationTests({
 }: DeploymentAgnosticFtrProviderContext) {
   describe('APM', function () {
     loadTestFile(require.resolve('./agent_explorer'));
+    loadTestFile(require.resolve('./custom_dashboards'));
   });
 }
